int changed to size_t in function definition

+++ cpp07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: upolat <upolat@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/06 17:07:48 by upolat            #+#    #+#             */
-/*   Updated: 2025/03/06 19:03:50 by upolat           ###   ########.fr       */
+/*   Updated: 2025/03/06 19:43:41 by upolat           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ template <class T> void print(T &x)
 	std::cout << x << std::endl;
 }
 
-template <class T> void iter(T *arr, int &len, void (*func)(T &elem))
+template <class T> void iter(T *arr, size_t &len, void (*func)(T &elem))
 {
-	for (int i = 0; i < len; i++)
+	for (size_t i = 0; i < len; i++)
 		(func)(arr[i]);
 }
